Add check to clean task
